services/search/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
pgvector
pydantic
sentence-transformers
transformers
torch
numpy
google-genai
